[*] Select: components separated into files.
