+++ primitives/src/select/components/mod.rs
@@ -1,0 +1,13 @@
+//! Component definitions for the select primitive.
+
+pub mod select;
+pub mod trigger;
+pub mod list;
+pub mod option;
+pub mod group;
+
+pub use select::{Select, SelectProps};
+pub use trigger::{SelectTrigger, SelectTriggerProps};
+pub use list::{SelectList, SelectListProps};
+pub use option::{SelectOption, SelectOptionProps, SelectItemIndicator, SelectItemIndicatorProps};
+pub use group::{SelectGroup, SelectGroupProps, SelectGroupLabel, SelectGroupLabelProps};
