+++ primitives/src/select/components/option.rs
@@ -1,0 +1,166 @@
+//! SelectOption and SelectItemIndicator component implementations.
+
+use crate::{
+    focus::use_focus_controlled_item,
+    use_effect, use_effect_cleanup, use_unique_id, use_id_or
+};
+use dioxus::html::input_data::MouseButton;
+use dioxus::prelude::*;
+use std::fmt::Display;
+
+use super::super::context::{
+    OptionState, SelectContext, SelectCursor, SelectOptionContext,
+};
+
+/// The props for the [`SelectOption`] component
+#[derive(Props, Clone, PartialEq)]
+pub struct SelectOptionProps<T: Clone + PartialEq + 'static = String> {
+    /// The value of the option
+    pub value: ReadOnlySignal<T>,
+
+    /// Optional display text (defaults to value.to_string())
+    #[props(default)]
+    pub display: Option<String>,
+
+    /// Whether the option is disabled
+    #[props(default)]
+    pub disabled: ReadOnlySignal<bool>,
+
+    /// Optional ID for the option
+    #[props(default)]
+    pub id: ReadOnlySignal<Option<String>>,
+
+    /// The index of the option in the list. This is used to define the focus order for keyboard navigation.
+    pub index: ReadOnlySignal<usize>,
+
+    /// Optional label for the option (for accessibility)
+    #[props(default)]
+    pub aria_label: Option<String>,
+
+    /// Optional description role for the option (for accessibility)
+    #[props(default)]
+    pub aria_roledescription: Option<String>,
+
+    #[props(extends = GlobalAttributes)]
+    attributes: Vec<Attribute>,
+
+    children: Element,
+}
+
+/// # SelectOption
+///
+/// An individual selectable option within a [`SelectList`] component. Each option represents
+/// a value that can be selected.
+///
+/// ## Value vs Display
+///
+/// - **`value`**: The programmatic value (e.g., `"apple"`, `"user_123"`) used internally
+/// - **`display`**: The user-facing text (e.g., `"Apple"`, `"John Doe"`) shown in the UI
+///
+/// If `display` is not provided, the `value` will be formatted and used as the display text.
+///
+/// This must be used inside a [`SelectList`] component.
+#[component]
+pub fn SelectOption<T: Display + PartialEq + Clone + 'static>(
+    props: SelectOptionProps<T>,
+) -> Element {
+    // Generate a unique ID for this option for accessibility
+    let option_id = use_unique_id();
+
+    // Use use_id_or to handle the ID
+    let id = use_id_or(option_id, props.id);
+
+    let index = props.index;
+    let value = props.value;
+    let display = use_memo(move || {
+        props
+            .display
+            .clone()
+            .unwrap_or_else(|| format!("{}", props.value.read()))
+    });
+
+    // Push this option to the context
+    let mut ctx: SelectContext<T> = use_context();
+    use_effect(move || {
+        let option_state = OptionState {
+            tab_index: index(),
+            value: value.cloned(),
+            display: display.read().to_string(),
+            id: id(),
+        };
+
+        // Add the option to the context's options
+        ctx.options.write().push(option_state);
+    });
+
+    use_effect_cleanup(move || {
+        ctx.options.write().retain(|opt| opt.id != *id.read());
+    });
+
+    let onmounted = use_focus_controlled_item(props.index);
+    let focused = move || ctx.focus_state.is_focused(index());
+    let disabled = ctx.disabled.cloned() || props.disabled.cloned();
+    let selected = use_memo(move || ctx.cursor.read().value == *props.value.read());
+
+    use_context_provider(|| SelectOptionContext {
+        selected: selected.into(),
+    });
+
+    rsx! {
+        div {
+            role: "option",
+            id,
+            tabindex: if focused() { "0" } else { "-1" },
+            onmounted,
+
+            // ARIA attributes
+            aria_selected: selected(),
+            aria_disabled: disabled,
+            aria_label: props.aria_label.clone(),
+            aria_roledescription: props.aria_roledescription.clone(),
+
+            onpointerdown: move |event| {
+                if !disabled && event.trigger_button() == Some(MouseButton::Primary) {
+                    ctx.set_value.call(Some(SelectCursor {
+                        value: props.value.read().clone(),
+                        display: display.read().to_string(),
+                    }));
+                    ctx.open.set(false);
+                }
+            },
+            onblur: move |_| {
+                if focused() {
+                    ctx.focus_state.blur();
+                    ctx.open.set(false);
+                }
+            },
+
+            ..props.attributes,
+            {props.children}
+        }
+    }
+}
+
+/// The props for the [`SelectItemIndicator`] component
+#[derive(Props, Clone, PartialEq)]
+pub struct SelectItemIndicatorProps {
+    /// The children to render inside the indicator
+    children: Element,
+}
+
+/// # SelectItemIndicator
+///
+/// The `SelectItemIndicator` component is used to render an indicator for a selected item within a [`SelectList`]. The
+/// children will only be rendered if the option is selected.
+///
+/// This must be used inside a [`SelectOption`] component.
+#[component]
+pub fn SelectItemIndicator(props: SelectItemIndicatorProps) -> Element {
+    let ctx: SelectOptionContext = use_context();
+    if !(ctx.selected)() {
+        return rsx! {};
+    }
+    rsx! {
+        {props.children}
+    }
+}
